+++ ref-exchange/src/account_deposit.rs
@@ -1,6 +1,7 @@
 //! Account deposit is information per user about their balances in the exchange.
 
 use std::collections::HashMap;
+use near_sdk::collections::UnorderedMap;
 
 use near_contract_standards::fungible_token::core_impl::ext_fungible_token;
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
@@ -9,7 +10,7 @@ use near_sdk::{
     assert_one_yocto, env, near_bindgen, 
     AccountId, Balance, PromiseResult, StorageUsage,
 };
-
+use crate::legacy::AccountV1;
 use crate::utils::{ext_self, GAS_FOR_FT_TRANSFER, GAS_FOR_RESOLVE_TRANSFER};
 use crate::*;
 
@@ -25,20 +26,33 @@ const U32_STORAGE: StorageUsage = 4;
 const ACC_ID_STORAGE: StorageUsage = 64;
 /// As a key, 4 bytes length would be added to the head
 const ACC_ID_AS_KEY_STORAGE: StorageUsage = ACC_ID_STORAGE + 4;
+const KEY_PREFIX_ACC: StorageUsage = 64;
 /// As a near_sdk::collection key, 1 byte for prefiex
 const ACC_ID_AS_CLT_KEY_STORAGE: StorageUsage = ACC_ID_AS_KEY_STORAGE + 1;
 
 // ACC_ID: the Contract accounts map key length
 // + VAccount enum: 1 byte
 // + U128_STORAGE: near_amount storage
+// + U32_STORAGE: legacy_tokens HashMap length
 // + U32_STORAGE: tokens HashMap length
 // + U64_STORAGE: storage_used
 pub const INIT_ACCOUNT_STORAGE: StorageUsage =
-    ACC_ID_AS_CLT_KEY_STORAGE + 1 + U128_STORAGE + U32_STORAGE + U64_STORAGE;
+    ACC_ID_AS_CLT_KEY_STORAGE + 1 + U128_STORAGE + U32_STORAGE + U32_STORAGE + U64_STORAGE;
 
 #[derive(BorshDeserialize, BorshSerialize)]
 pub enum VAccount {
+    V1(AccountV1),
     Current(Account),
+}
+
+impl VAccount {
+    /// Upgrades from other versions to the currently used version.
+    pub fn into_current(self, account_id: &AccountId) -> Account {
+        match self {
+            VAccount::Current(account) => account,
+            VAccount::V1(account) => account.into_current(account_id),
+        }
+    }
 }
 
 impl From<Account> for VAccount {
@@ -47,52 +61,109 @@ impl From<Account> for VAccount {
     }
 }
 
-impl From<VAccount> for Account {
-    fn from(v_account: VAccount) -> Self {
-        match v_account {
-            VAccount::Current(account) => {account},
-        }
-    }
-}
-
 
 /// Account deposits information and storage cost.
-#[derive(BorshSerialize, BorshDeserialize, Default, Clone)]
+#[derive(BorshSerialize, BorshDeserialize)]
 pub struct Account {
     /// Native NEAR amount sent to the exchange.
     /// Used for storage right now, but in future can be used for trading as well.
     pub near_amount: Balance,
     /// Amounts of various tokens deposited to this account.
-    pub tokens: HashMap<AccountId, Balance>,
+    pub legacy_tokens: HashMap<AccountId, Balance>,
+    pub tokens: UnorderedMap<AccountId, Balance>,
     pub storage_used: StorageUsage,
 }
 
-
-
-
 impl Account {
+    pub fn new(account_id: &AccountId) -> Self {
+        Account {
+            near_amount: 0,
+            legacy_tokens: HashMap::new(),
+            tokens: UnorderedMap::new(StorageKey::AccountTokens {
+                account_id: account_id.clone(),
+            }),
+            storage_used: 0,
+        }
+    }
+
+    pub fn get_balance(&self, token_id: &AccountId) -> Option<Balance> {
+        if let Some(token_balance) = self.tokens.get(token_id) {
+            Some(token_balance)
+        } else if let Some(legacy_token_balance) = self.legacy_tokens.get(token_id) {
+            Some(*legacy_token_balance)
+        } else {
+            None
+        }
+    }
+
+    pub fn get_tokens(&self) -> Vec<AccountId> {
+        let mut a: Vec<AccountId> = self.tokens.keys().collect();
+        let b: Vec<AccountId> = self.legacy_tokens
+            .keys()
+            .cloned()
+            .collect();
+        a.extend(b);
+        a
+    }
+
+    /// Deposit amount to the balance of given token,
+    /// if given token not register and not enough storage, deposit fails 
+    pub(crate) fn deposit_with_storage_check(&mut self, token: &AccountId, amount: Balance) -> bool { 
+        if let Some(balance) = self.tokens.get(token) {
+            // token has been registered, just add without storage check, 
+            let new_balance = balance + amount;
+            self.tokens.insert(token, &new_balance);
+            true
+        } else if let Some(x) = self.legacy_tokens.get_mut(token) {
+            // token has been registered, just add without storage check
+            *x += amount;
+            true
+        } else {
+            // check storage after insert, if fail should unregister the token
+            self.tokens.insert(token, &(amount));
+            if self.storage_usage() <= self.near_amount {
+                true
+            } else {
+                self.tokens.remove(token);
+                false
+            }
+        }
+    }
+
     /// Deposit amount to the balance of given token.
     pub(crate) fn deposit(&mut self, token: &AccountId, amount: Balance) {
-        if let Some(x) = self.tokens.get_mut(token) {
-            *x = *x + amount;
+        if let Some(x) = self.legacy_tokens.remove(token) {
+            // need convert to tokens
+            self.tokens.insert(token, &(amount + x));
+        } else if let Some(x) = self.tokens.get(token) {
+            self.tokens.insert(token, &(amount + x));
         } else {
-            self.tokens.insert(token.clone(), amount);
+            self.tokens.insert(token, &amount);
         }
     }
 
     /// Withdraw amount of `token` from the internal balance.
     /// Panics if `amount` is bigger than the current balance.
     pub(crate) fn withdraw(&mut self, token: &AccountId, amount: Balance) {
-        let value = *self.tokens.get(token).expect(ERR21_TOKEN_NOT_REG);
-        assert!(value >= amount, "{}", ERR22_NOT_ENOUGH_TOKENS);
-        self.tokens.insert(token.clone(), value - amount);
+        if let Some(x) = self.legacy_tokens.remove(token) {
+            // need convert to 
+            assert!(x >= amount, "{}", ERR22_NOT_ENOUGH_TOKENS);
+            self.tokens.insert(token, &(x - amount));
+        } else if let Some(x) = self.tokens.get(token) {
+            assert!(x >= amount, "{}", ERR22_NOT_ENOUGH_TOKENS);
+            self.tokens.insert(token, &(x - amount));
+        } else {
+            env::panic(ERR21_TOKEN_NOT_REG.as_bytes());
+        }
     }
 
     // [AUDIT_01]
     /// Returns amount of $NEAR necessary to cover storage used by this data structure.
     pub fn storage_usage(&self) -> Balance {
         (INIT_ACCOUNT_STORAGE + 
-            self.tokens.len() as u64 * (ACC_ID_AS_KEY_STORAGE + U128_STORAGE)) as u128
+            self.legacy_tokens.len() as u64 * (ACC_ID_AS_KEY_STORAGE + U128_STORAGE) + 
+            self.tokens.len() as u64 * (KEY_PREFIX_ACC + ACC_ID_AS_KEY_STORAGE + U128_STORAGE)
+        ) as u128
             * env::storage_byte_cost()
     }
 
@@ -122,12 +193,11 @@ impl Account {
     }
 
     /// Registers given token and set balance to 0.
-    /// Fails if not enough amount to cover new storage usage.
     pub(crate) fn register(&mut self, token_ids: &Vec<ValidAccountId>) {
         for token_id in token_ids {
             let t = token_id.as_ref();
-            if !self.tokens.contains_key(t) {
-                self.tokens.insert(t.clone(), 0);
+            if self.get_balance(t).is_none() {
+                self.tokens.insert(t, &0);
             }
         }
     }
@@ -135,6 +205,8 @@ impl Account {
     /// Unregisters `token_id` from this account balance.
     /// Panics if the `token_id` balance is not 0.
     pub(crate) fn unregister(&mut self, token_id: &AccountId) {
+        let amount = self.legacy_tokens.remove(token_id).unwrap_or_default();
+        assert_eq!(amount, 0, "{}", ERR24_NON_ZERO_TOKEN_BALANCE);
         let amount = self.tokens.remove(token_id).unwrap_or_default();
         assert_eq!(amount, 0, "{}", ERR24_NON_ZERO_TOKEN_BALANCE);
     }
@@ -149,9 +221,9 @@ impl Contract {
         assert_one_yocto();
         self.assert_contract_running();
         let sender_id = env::predecessor_account_id();
-        let mut deposits = self.internal_unwrap_account(&sender_id);
-        deposits.register(&token_ids);
-        self.internal_save_account(&sender_id, deposits);
+        let mut account = self.internal_unwrap_account(&sender_id);
+        account.register(&token_ids);
+        self.internal_save_account(&sender_id, account);
     }
 
     /// Unregister given token from user's account deposit.
@@ -161,11 +233,11 @@ impl Contract {
         assert_one_yocto();
         self.assert_contract_running();
         let sender_id = env::predecessor_account_id();
-        let mut deposits = self.internal_unwrap_account(&sender_id);
+        let mut account = self.internal_unwrap_account(&sender_id);
         for token_id in token_ids {
-            deposits.unregister(token_id.as_ref());
+            account.unregister(token_id.as_ref());
         }
-        self.internal_save_account(&sender_id, deposits);
+        self.internal_save_account(&sender_id, account);
     }
 
     /// Withdraws given token from the deposits of given user.
@@ -182,14 +254,15 @@ impl Contract {
         self.assert_contract_running();
         let token_id: AccountId = token_id.into();
         let amount: u128 = amount.into();
+        assert!(amount > 0, "{}", ERR29_ILLEGAL_WITHDRAW_AMOUNT);
         let sender_id = env::predecessor_account_id();
-        let mut deposits = self.internal_unwrap_account(&sender_id);
+        let mut account = self.internal_unwrap_account(&sender_id);
         // Note: subtraction and deregistration will be reverted if the promise fails.
-        deposits.withdraw(&token_id, amount);
+        account.withdraw(&token_id, amount);
         if unregister == Some(true) {
-            deposits.unregister(&token_id);
+            account.unregister(&token_id);
         }
-        self.internal_save_account(&sender_id, deposits);
+        self.internal_save_account(&sender_id, account);
         self.internal_send_tokens(&sender_id, &token_id, amount)
     }
 
@@ -210,22 +283,38 @@ impl Contract {
             PromiseResult::NotReady => unreachable!(),
             PromiseResult::Successful(_) => {}
             PromiseResult::Failed => {
-                // This reverts the changes from withdraw function. If account doesn't exit, deposits to the owner's account.
-                if let Some(va) = self.accounts.get(&sender_id) {
-                    let mut account: Account = va.into();
-                    account.deposit(&token_id, amount.0);
-                    self.internal_save_account(&sender_id, account);
+                // This reverts the changes from withdraw function.
+                // If account doesn't exit, deposits to the owner's account as lostfound.
+                let mut failed = false;
+                if let Some(mut account) = self.internal_get_account(&sender_id) {
+                    if account.deposit_with_storage_check(&token_id, amount.0) {
+                        // cause storage already checked, here can directly save
+                        self.accounts.insert(&sender_id, &account.into());
+                    } else {
+                        // we can ensure that internal_get_account here would NOT cause a version upgrade, 
+                        // cause it is callback, the account must be the current version or non-exist,
+                        // so, here we can just leave it without insert, won't cause storage collection inconsistency.
+                        env::log(
+                            format!(
+                                "Account {} has not enough storage. Depositing to owner.",
+                                sender_id
+                            )
+                            .as_bytes(),
+                        );
+                        failed = true;
+                    }
                 } else {
                     env::log(
                         format!(
-                            "Account {} is not registered or not enough storage. Depositing to owner.",
+                            "Account {} is not registered. Depositing to owner.",
                             sender_id
                         )
                         .as_bytes(),
                     );
-                    let mut owner_account = self.internal_unwrap_account(&self.owner_id);
-                    owner_account.deposit(&token_id, amount.0);
-                    self.internal_save_account(&self.owner_id.clone(), owner_account);
+                    failed = true;
+                }
+                if failed {
+                    self.internal_lostfound(&token_id, amount.0);
                 }
             }
         };
@@ -241,6 +330,20 @@ impl Contract {
         self.accounts.insert(&account_id, &account.into());
     }
 
+    /// save token to owner account as lostfound, no need to care about storage
+    /// only global whitelisted token can be stored in lost-found
+    pub(crate) fn internal_lostfound(&mut self, token_id: &AccountId, amount: u128) {
+        if self.whitelisted_tokens.contains(token_id) {
+            let mut lostfound = self.internal_unwrap_or_default_account(&self.owner_id);
+            lostfound.deposit(token_id, amount);
+            self.accounts.insert(&self.owner_id, &lostfound.into());
+        } else {
+            env::panic("ERR: non-whitelisted token can NOT deposit into lost-found.".as_bytes());
+        }
+        
+    }
+    
+
     /// Registers account in deposited amounts with given amount of $NEAR.
     /// If account already exists, adds amount to it.
     /// This should be used when it's known that storage is prepaid.
@@ -254,6 +357,7 @@ impl Contract {
     pub(crate) fn internal_storage_withdraw(&mut self, account_id: &AccountId, amount: Balance) -> u128 {
         let mut account = self.internal_unwrap_account(&account_id);
         let available = account.storage_available();
+        assert!(available > 0, "ERR_NO_STORAGE_CAN_WITHDRAW");
         let mut withdraw_amount = amount;
         if amount == 0 {
             withdraw_amount = available;
@@ -274,7 +378,8 @@ impl Contract {
     ) {
         let mut account = self.internal_unwrap_account(sender_id);
         assert!(
-            self.whitelisted_tokens.contains(token_id) || account.tokens.contains_key(token_id),
+            self.whitelisted_tokens.contains(token_id) 
+                || account.get_balance(token_id).is_some(),
             "{}",
             ERR12_TOKEN_NOT_WHITELISTED
         );
@@ -282,18 +387,20 @@ impl Contract {
         self.internal_save_account(&sender_id, account);
     }
 
-    pub fn internal_unwrap_account(&self, account_id: &AccountId) -> Account {
+    pub fn internal_get_account(&self, account_id: &AccountId) -> Option<Account> {
         self.accounts
             .get(account_id)
+            .map(|va| va.into_current(account_id))
+    }
+
+    pub fn internal_unwrap_account(&self, account_id: &AccountId) -> Account {
+        self.internal_get_account(account_id)
             .expect(errors::ERR10_ACC_NOT_REGISTERED)
-            .into()
     }
 
     pub fn internal_unwrap_or_default_account(&self, account_id: &AccountId) -> Account {
-        self.accounts
-            .get(account_id)
-            .unwrap_or(Account::default().into())
-            .into()
+        self.internal_get_account(account_id)
+            .unwrap_or_else(|| Account::new(account_id))
     }
 
     /// Returns current balance of given token for given user. If there is nothing recorded, returns 0.
@@ -302,10 +409,9 @@ impl Contract {
         sender_id: &AccountId,
         token_id: &AccountId,
     ) -> Balance {
-        self.accounts
-            .get(sender_id)
-            .and_then(|va| {let a: Account = va.into(); a.tokens.get(token_id).cloned()})
-            .unwrap_or_default()
+        self.internal_get_account(sender_id)
+            .and_then(|x| x.get_balance(token_id))
+            .unwrap_or(0)
     }
 
     /// Sends given amount to given user and if it fails, returns it back to user's balance.