+++ ref-exchange/src/errors.rs
@@ -18,6 +18,8 @@ pub const ERR25_CALLBACK_POST_WITHDRAW_INVALID: &str =
 // pub const ERR26_ACCESS_KEY_NOT_ALLOWED: &str = "E26: access key not allowed";
 pub const ERR27_DEPOSIT_NEEDED: &str = 
     "E27: attach 1yN to swap tokens not in whitelist";
+pub const ERR28_WRONG_MSG_FORMAT: &str = "E28: Illegal msg in ft_transfer_call";
+pub const ERR29_ILLEGAL_WITHDRAW_AMOUNT: &str = "E29: Illegal withdraw amount";
 
 // Liquidity operations.