+++ ref-exchange/src/storage_impl.rs
@@ -59,9 +59,8 @@ impl StorageManagement for Contract {
         assert_one_yocto();
         self.assert_contract_running();
         let account_id = env::predecessor_account_id();
-        if let Some(account_deposit) = self.accounts.get(&account_id) {
+        if let Some(account_deposit) = self.internal_get_account(&account_id) {
             // TODO: figure out force option logic.
-            let account_deposit: Account = account_deposit.into();
             assert!(
                 account_deposit.tokens.is_empty(),
                 "ERR_STORAGE_UNREGISTER_TOKENS_NOT_EMPTY"
@@ -82,11 +81,9 @@ impl StorageManagement for Contract {
     }
 
     fn storage_balance_of(&self, account_id: ValidAccountId) -> Option<StorageBalance> {
-        self.accounts
-            .get(account_id.as_ref())
-            .map(|deposits| 
+        self.internal_get_account(account_id.as_ref())
+            .map(|account| 
                 { 
-                    let account: Account = deposits.into(); 
                     StorageBalance {
                         total: U128(account.near_amount),
                         available: U128(account.storage_available()),