+++ ref-exchange/tests/test_migrate.rs
@@ -1,6 +1,6 @@
 use std::convert::TryFrom;
 
-use near_sdk::json_types::ValidAccountId;
+use near_sdk::json_types::{ValidAccountId};
 use near_sdk_sim::{deploy, init_simulator, to_yocto};
 
 use ref_exchange::{ContractContract as Exchange, RunningState};
@@ -9,7 +9,7 @@ use crate::common::utils::*;
 pub mod common;
 
 near_sdk_sim::lazy_static_include::lazy_static_include_bytes! {
-    PREV_EXCHANGE_WASM_BYTES => "../res/ref_exchange_101.wasm",
+    PREV_EXCHANGE_WASM_BYTES => "../res/ref_exchange_102.wasm",
     EXCHANGE_WASM_BYTES => "../res/ref_exchange_release.wasm",
 }
 
@@ -46,7 +46,7 @@ fn test_upgrade() {
     .assert_success();
     let metadata = get_metadata(&pool);
     // println!("{:#?}", metadata);
-    assert_eq!(metadata.version, "1.1.0".to_string());
+    assert_eq!(metadata.version, "1.2.0".to_string());
     assert_eq!(metadata.state, RunningState::Running);
 
     // Upgrade to the same code migration is skipped.